backend/requirements.txt
fastapi>=0.115.0,<1
uvicorn[standard]>=0.32.0,<1
anthropic>=0.40.0,<1
python-multipart>=0.0.12,<1
supabase>=2.10.0,<3
python-dotenv>=1.0.0,<2
pydantic>=2.10.0,<3
httpx>=0.27.0,<1
rich>=13.0.0,<14
